pyproject.toml
[project]
name = "ai-hr-assistant"
version = "0.1.0"
description = "AI-powered HR assistant for resume authenticity scanning and JD matching"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "python-docx>=1.1.0",
    "docx2txt>=0.8",
    "docxtpl>=0.16.7",
    "pytesseract>=0.3.10",
    "opencv-python>=4.8.1.78",
    "pdfplumber>=0.10.3",
    "pymupdf>=1.24.9",
    "Pillow>=10.1.0",
    "spacy>=3.7.2",
    "nltk>=3.8.1",
    "jinja2>=3.1.2",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "aiofiles>=23.2.1",
    "python-magic>=0.4.27",
    "numpy>=1.24.3",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.3",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
    "celery>=5.3.4",
    "redis>=5.0.1",
    "PyJWT>=2.8.0",
    "bcrypt>=4.1.1",
    "email-validator>=2.1.0",
    "passlib>=1.7.4",
    "sqlalchemy[asyncio]>=2.0.23",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
    "sendgrid>=6.11.0",
    "PyPDF2>=3.0.1",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
